Split previous combined dialog specs into per dialog versions
